% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{average_minimum_flow}
\alias{average_minimum_flow}
\title{Average Minimum Flow}
\usage{
average_minimum_flow(ts, n = 1, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}. Assumes a regular timeseries.}

\item{n}{The moving-window size.}

\item{which}{Logical: If \code{TRUE}, return the index location of the 
n-average minimum flow instead of the flow magnitude.}

\item{return}{The n-average minimum flow.}
}
\description{
Compute the n-average minimum flow.
}
\examples{
data(siouxcity)
average_minimum_flow(siouxcity['2009'], 1)
average_minimum_flow(siouxcity['2009'], 7)
average_minimum_flow(siouxcity['2009'], 90, which = TRUE)

}

