% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{baseflow_index}
\alias{baseflow_index}
\title{Baseflow Index}
\usage{
baseflow_index(ts)
}
\arguments{
\item{ts}{A time series of class \code{xts}. Assumes a regular timeseries.}
}
\value{
The baseflow index.
}
\description{
Compute the baseflow index.
}
\details{
The baseflow index is defined as the minimum 7-day average flow
  normalized by the mean flow. The baseflow index is typically computed for
  a single year of record.
}

