% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rva.r
\name{build_RVA_categories}
\alias{build_RVA_categories}
\title{Define RVA Category Boundaries}
\usage{
build_RVA_categories(pre, boundaries, parametric = c(FALSE, FALSE),
  na.rm = FALSE)
}
\arguments{
\item{pre}{The pre-impact IHA statistics, i.e. the output of 
\code{IHA(..., keep.raw = TRUE)}.}

\item{boundaries}{a two-element numeric vector of cutoffs that define the
three RVA categories.}

\item{parametric}{A logical vector with 1 or 2 elements: Compute lower and
upper boundaries for RVA categories based on parametric (mean) or 
non-parametric (median) analysis.}

\item{na.rm}{Logical: Remove NA values when computing means/medians.}
}
\value{
A 4-column dataframe containing the upper and lower bounds of the 
  RVA categories for each parameter.
}
\description{
Compute the cutoff values for specifying the three categories used in the
RVA approach.
}

