% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.r
\name{check_EFC_thresholds}
\alias{check_EFC_thresholds}
\title{Check EFC Thresholds}
\usage{
check_EFC_thresholds(efcthresh, ts)
}
\arguments{
\item{efcthresh}{The EFC thresholds, as defined by e.g. 
\code{build_EFC_thresholds(...)}.}

\item{ts}{The time series that the EFC thresholds will be applied to.}
}
\value{
A dataframe of 3 columns:
}
\description{
Check the suitability of EFC thresholds. If thresholds result in classes 
outside the range of values in the pre-impact dataset for any parameter, 
a warning will be issued.
}

