% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rva.r
\name{check_RVA_categories}
\alias{check_RVA_categories}
\title{Check RVA Categories}
\usage{
check_RVA_categories(rvacat, pre)
}
\arguments{
\item{rvacat}{The RVA categories, as defined by e.g. RVA_categories(pre).}

\item{pre}{The pre-impact IHA statistics used to define the RVA categories, 
i.e. the output of \code{IHA(...)}.}
}
\value{
A dataframe of 3 columns:
}
\description{
Check the suitability of RVA categories. If category cutoffs are defined 
outside the range of values in the pre-impact dataset for any parameter, 
a warning will be issued.
}

