% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{circ_stat}
\alias{circ_stat}
\title{Circular Statistics for IHA}
\usage{
circ_stat(v, statfun)
}
\arguments{
\item{v}{Vector of Julian dates}

\item{statfun}{the function to be applied, e.g. \code{mean}, \code{median}, 
\code{sd}, etc.}
}
\value{
The result of the circular statistic calculation.
}
\description{
Compute IHA circular statistics for working with Julian dates
}
\details{
The algorithm works by assigning values in \code{v} to four bins
  of equal width. Julian dates in certain bins are then adjusted by +/- 366 
  depending on the which bin contains the most elements of \code{v}. After 
  \code{statfun} is applied, the result is adjusted by +/- 366 as necessary 
  to report a valid Julian date. The function will warn the user if more 
  than 10% of the elements of \code{v} fall within the bin six months away 
  from the primary bin; in cases where the data exhibits considerable 
  scatter among Julian dates, the circular statistics will be less
  informative and should be interpreted with caution.
}

