% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{compareIHA}
\alias{compareIHA}
\title{Compare IHA Results}
\usage{
compareIHA(pre, post)
}
\arguments{
\item{pre}{The pre-impact IHA statistics, i.e. the output of 
\code{IHA(...)}.}

\item{post}{The post-impact IHA statistics.}
}
\value{
A dataframe containing the IHA deviation factors.
}
\description{
Compare the results of two IHA analyses, e.g. pre- and post-impact periods.
}
\examples{
data(siouxcity)
pre = IHA(siouxcity['2004/2009'], ut = 32000, lt = 12000)
post = IHA(siouxcity['2010/2014'], ut = 32000, lt = 12000)
compareIHA(pre, post)

}

