% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{confint.IHA}
\alias{confint.IHA}
\title{Confidence Intervals for IHA Central Tendencies}
\usage{
\method{confint}{IHA}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{IHA}.}

\item{parm}{A specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, 
all parameters are considered.}

\item{level}{The confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence 
  limits for each parameter. These will be labelled as (1 - level)/2 and 
  1 - (1 - level)/2 in % (by default 2.5% and 97.5%).
}
\description{
Computes confidence intervals for parametric and non-parametric central 
tendencies computed by IHA.
}
\details{
Parametric confidence intervals are computed using \code{t-test}.
  Non-parametric confidence intervals are based on parameter value rankings.
}
\examples{
data(siouxcity)
res = IHA(siouxcity['2009/2011'], ut = 32000, lt = 12000)
summary(res)
confint(res)

}

