% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.r
\name{EFC}
\alias{EFC}
\title{IHA Environmental Flow Components}
\usage{
EFC(ts, yearstart = "01-01", yearend = "12-31", method = c("standard",
  "advanced"), thresholds = list())
}
\description{
Compute the IHA Environmental Flow Components (EFCs).
}

