% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\docType{package}
\name{flowregime-package}
\alias{flowregime}
\alias{flowregime-package}
\title{Environmental Flow Regime Components}
\description{
A package for computing environmental flow regime components
from hydrologic time series. See the vignette to get started.
}

