% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{get_jd}
\alias{get_jd}
\title{IHA Julian Day}
\usage{
get_jd(d)
}
\arguments{
\item{d}{Date or POSIX*t object}
}
\value{
Integer Julian date.
}
\description{
Compute Julian Day as per IHA v7.
}
\details{
IHA defines Jan 1 as JD 1 and Dec 31 as JD 366 for both leap years
  and non-leap years. In non-leap years, JD 60 (Feb 29) is skipped, so that
  e.g. Feb 28 2003 is JD 59 and Mar 01 2003 is JD 61.
}

