% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{IHAsignif}
\alias{IHAsignif}
\title{Significance Counts for IHA Comparison}
\usage{
IHAsignif(pre, post)
}
\description{
Computes significance counts for non-parametric comparison of IHA objects
}

