% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{low_flow_duration}
\alias{low_flow_duration}
\title{Low Flow Duration}
\usage{
low_flow_duration(ts, lt)
}
\arguments{
\item{ts}{A time series, e.g. of class \code{zoo}.}

\item{lt}{The lower flow threshold used to compute duration.}
}
\value{
The duration of the longest period where flow is at
  or below the threshold, in a format dependent on the value
  of \code{index(ts)}.
}
\description{
Compute the longest continuous period during which flow is
  at or below a given threshold.
}
\examples{
testts = zoo(c(sort(rlnorm(1000)), rev(sort(rlnorm(489)))),
  order.by = seq(as.POSIXct("2010-01-01 00:00:00"),
  as.POSIXct("2010-02-01 00:00:00"), length.out = 1489),
  frequency=2*24*365)
lower = quantile(testts, 0.25)
low_flow_duration(testts, lower)
}

