% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{mean_high_pulse_duration}
\alias{mean_high_pulse_duration}
\title{Mean Duration of High Flow Pulses}
\usage{
mean_high_pulse_duration(ts, ut)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{ut}{The upper flow threshold above which to identify pulses.}
}
\value{
The mean duration of high flow pulses.
}
\description{
Compute the average duration of high flow pulses above a threshold.
}

