% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{number_of_high_pulses}
\alias{number_of_high_pulses}
\title{Number of High Flow Pulses}
\usage{
number_of_high_pulses(ts, ut, min.dur = NA, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{ut}{The upper flow threshold above which to identify peaks.}

\item{min.dur}{The minimum duration required for a pulse to be counted.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the 
peaks instead of the count.}
}
\value{
The number of peaks above the threshold.
}
\description{
Compute the number of high flow pulses (peaks) above a threshold.
}
\examples{
data(siouxcity)
number_of_high_pulses(siouxcity['2009'], ut = 32000)
number_of_high_pulses(siouxcity['2009'], ut = 32000, which = TRUE)

}

