% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{number_of_low_pulses}
\alias{number_of_low_pulses}
\title{Number of Low Flow Pulses}
\usage{
number_of_low_pulses(ts, lt, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{lt}{The lower flow threshold below which to identify pulses.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the 
pulses instead of the count.}
}
\value{
The number of pulses below the threshold.
}
\description{
Compute the number of low flow pulses below a threshold.
}
\examples{
data(siouxcity)
number_of_low_pulses(siouxcity['2009'], lt = 12000)
number_of_low_pulses(siouxcity['2009'], lt = 12000, which = TRUE)

}

