% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{number_of_no_flow_days}
\alias{number_of_no_flow_days}
\title{Number of No-Flow Days}
\usage{
number_of_no_flow_days(ts, tol = 0)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{tol}{A tolerance factor. Flows less than \code{tol} are considered to 
be essentially zero.}
}
\value{
The number of no-flow days.
}
\description{
Compute the number of days for which flow is zero (to some tolerance).
}
\examples{
data(siouxcity)
number_of_no_flow_days(siouxcity['2009'])
number_of_no_flow_days(siouxcity['2009'], tol = 0.1)

}

