% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{number_of_reversals}
\alias{number_of_reversals}
\title{Number Of Flow Reversals}
\usage{
number_of_reversals(ts, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}. Assumes a regular timeseries.}

\item{which}{Logical: If \code{TRUE}, return the index location of the 
flow reversals instead of the total number of reversals.}

\item{return}{The total number of flow reversals.}
}
\description{
Compute the number of times that the flow rate of change reverses.
}
\examples{
data(siouxcity)
number_of_reversals(siouxcity['2009-01'])
number_of_reversals(siouxcity['2009-01'], which = TRUE)

}

