% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\docType{data}
\name{roanokerapids}
\alias{roanokerapids}
\title{Roanoke River Flows at Roanoke Rapids, NC}
\format{An \code{xts} object.}
\usage{
data(roanokerapids)
}
\description{
Roanoke River daily discharge recorded at USGS gauge 02080500 in 
  Roanoke Rapids, NC from 1913-01-01 to 1991-12-31.
}
\keyword{datasets}

