% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rva.r
\name{RVA}
\alias{RVA}
\title{IHA Range of Variability Approach}
\usage{
RVA(pre, post, boundaries, parametric = TRUE)
}
\arguments{
\item{pre}{The pre-impact IHA statistics, i.e. the output of 
\code{IHA(..., keep.raw = TRUE)}.}

\item{post}{The post-impact IHA statistics.}

\item{parametric}{Logical: perform parametric (mean) or non-parametric
(median) analysis.}
}
\value{
A dataframe.
}
\description{
Compute the IHA Range of Variability.
}

