% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{separate_flow}
\alias{separate_flow}
\title{Separate Flow Data}
\usage{
separate_flow(ts, filter_parameter = 0.925, passes = 3)
}
\arguments{
\item{ts}{A time series of  of class \code{xts}.}

\item{filter_parameter}{Filter parameter. Default value is that
recommended by Nathan and McMahon (1990).}

\item{passes}{The number of times to pass the filter over the data.}
}
\value{
An \code{xts} object containing the columns "baseflow" and
  "quickflow".
}
\description{
Separate flow data into baseflow and quickflow.
}
\details{
This function is basically a \code{zoo}/\code{xts} wrapper for
  \code{EcoHydRology::BaseflowSeparation}.
}
\seealso{
\code{\link[EcoHydRology]{BaseflowSeparation}}.
}

