% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\docType{data}
\name{siouxcity}
\alias{siouxcity}
\title{Missouri River Flows at Sioux City, IA}
\format{An \code{xts} object.}
\usage{
data(siouxcity)
}
\description{
Missouri River daily discharge recorded at USGS gauge 06486000 in
  Sioux City, IA from 2004-01-01 to 2015-05-25.
}
\keyword{datasets}

