% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{time_to_recede}
\alias{time_to_recede}
\title{Time To Recede}
\usage{
time_to_recede(ts, lt, ut)
}
\arguments{
\item{ts}{A time series, e.g. of class \code{zoo}.}

\item{lt}{The lower flow threshold for which the time to recede is
computed. If omitted, the minimum flow of the series is used.}

\item{ut}{The upper flow threshold from which to start computing
the time to recede. If omitted, the maximum flow of the series is
used.}
}
\value{
the time to recede, in a format dependent on the value of
  \code{index(ts)}.
}
\description{
Compute the time for flow to fall down to or below a given threshold.
}
\examples{
testts = zoo(c(sort(rlnorm(1000)), rev(sort(rlnorm(489)))),
  order.by = seq(as.POSIXct("2010-01-01 00:00:00"),
  as.POSIXct("2010-02-01 00:00:00"), length.out = 1489),
  frequency=2*24*365)
lower = quantile(testts, 0.25)
upper= quantile(testts, 0.75)
time_to_recede(testts, lower, upper)
}

