% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{total_time_above_threshold}
\alias{total_time_above_threshold}
\title{Total Time Above Threshold}
\usage{
total_time_above_threshold(ts, ut)
}
\arguments{
\item{ts}{A time series, e.g. of class \code{zoo}.}

\item{ut}{The upper flow threshold.}
}
\value{
The total amount of time that flow is at or
  above the threshold, in a format dependent on the
  value of \code{index(ts)}.
}
\description{
Compute the total amount of time that flow is
  at or above a given threshold.
}
\examples{
testts = zooreg(c(sort(rlnorm(1000)), rev(sort(rlnorm(489)))),
  order.by = seq(as.POSIXct("2010-01-01 00:00:00"),
  as.POSIXct("2010-02-01 00:00:00"), length.out = 1489),
  frequency=2*24*365)
upper = quantile(testts, 0.75)
total_time_above_threshold(testts, upper)
}

