% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{compareIHA}
\alias{compareIHA}
\title{Compare IHA Results}
\usage{
compareIHA(pre, post)
}
\arguments{
\item{pre}{The pre-impact IHA statistics, i.e. the output of 
\code{IHA(..., keep.raw = TRUE)}.}

\item{post}{The post-impact IHA statistics.}

\item{as.percent}{If \code{TRUE}, return the differences as a relative 
percent difference (\code{(post - pre)/pre}). Otherwise, return the 
magnitude of difference.}

\item{cl}{Logical: If \code{TRUE}, compute confidence limits for each 
parameter.}
}
\value{
A dataframe containing either the magnitude of difference or 
  relative percent difference of each parameter, and (optionally) the upper
  and lower confidence intervals.
}
\description{
Compare the results of two IHA analyses, e.g. pre- and post-impact periods.
}
\examples{
data(siouxcity)
pre = IHA(siouxcity['2004/2009'], ut = 32000, lt = 12000)
post = IHA(siouxcity['2010/2014'], ut = 32000, lt = 12000)
compareIHA(pre, post)
compareIHA(pre, post, as.percent = TRUE, cl = TRUE)

}

