% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.r
\name{EFC}
\alias{EFC}
\title{IHA Environmental Flow Components}
\usage{
EFC(ts, method = c("standard", "advanced"), thresholds)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{method}{The method used to compute environmental flow components:
may be either 'standard' or 'advanced'.}

\item{thresholds}{A list containing all or some of the following elements:}
}
\value{
A copy of \code{ts} with an additional column containing EFC labels.
}
\description{
Compute the IHA Environmental Flow Components (EFCs).
}
\examples{
data(roanokerapids)
EFC(roanokerapids)
EFC(roanokerapids, method = "advanced")

}

