% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.r
\name{EFC}
\alias{EFC}
\title{IHA Environmental Flow Components}
\usage{
EFC(ts, yearstart = "01-01", yearend = "12-31", method = c("standard",
  "advanced"), thresholds = list())
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{yearstart}{A character vector specifying the month and day signifying
the start of the water year, in the format "mm-dd". Default is "01-01".}

\item{yearend}{A character vector specifying the month and day signifying
the end of the water year, in the format "mm-dd". Default is "12-31".}

\item{method}{The method used to compute environmental flow components:
may be either 'standard' or 'advanced'.}

\item{thresholds}{A list containing all or some of the following elements:}
}
\value{
An \code{xts} time series of EFC factors.
}
\description{
Compute the IHA Environmental Flow Components (EFCs).
}

