% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{high_flow_duration}
\alias{high_flow_duration}
\title{High Flow Duration}
\usage{
high_flow_duration(ts, ut, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{ut}{The upper flow threshold used to compute duration.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the 
duration instead of the count.}
}
\value{
The duration of the longest period where flow is at
  or above the threshold, in a format dependent on the value 
  of \code{index(ts)}.
}
\description{
Compute the longest continuous period during which flow is 
  at or above a given threshold.
}
\examples{
data(siouxcity)
high_flow_duration(siouxcity['2011'], 70000)
high_flow_duration(siouxcity['2011'], 70000, which = TRUE)

}

