% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{IHA}
\alias{IHA}
\title{Indicators of Hydrologic Alteration}
\usage{
IHA(ts, yearstart = "01-01", yearend = "12-31", groups = 1:5, ut, lt,
  stats = TRUE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{yearstart}{A character vector specifying the month and day signifying
the start of the water year, in the format "mm-dd". Default is "01-01".}

\item{yearend}{A character vector specifying the month and day signifying
the end of the water year, in the format "mm-dd". Default is "12-31".}

\item{groups}{The parameter groups to compute. Computes all groups (1-5) by
default.}

\item{ut}{The upper flow threshold for identifying high flow pulses. IHA 
recommends the 75th flow percentile for pre-impact conditions.}

\item{lt}{The lower flow threshold for identifying low flow pulses. IHA 
recommends the 25th flow percentile for pre-impact conditions.}

\item{stats}{Logical: If \code{TRUE}, return the inter-annual statistics 
of each parameter. Otherwise, return the hydrologic attributes for each 
year of record.}
}
\value{
A 3-column dataframe. If \code{stat = TRUE}, the dataframe contains 
  the parameter names, the central tendencies and the measures of 
  dispersion. If \code{stat = FALSE}, the dataframe contains the parameter 
  names, the value of the attribute and the year of record (YoR) for which 
  the attribute was computed.
}
\description{
Compute the Indicators of Hydrologic Alteration for a daily time series.
}
\details{
The IHA method requires a regular time series with no missing 
  values.
}
\examples{
data(siouxcity)
IHA(siouxcity['2009/2011'], ut = 32000, lt = 12000)
IHA(siouxcity['2009/2011'], ut = 32000, lt = 12000, stats = FALSE)
IHA(siouxcity['2009-10-01/2011-09-30'], yearstart = "10-01", 
  yearend = "09-30", ut = 32000, lt = 12000)

}
\references{
Richter, B. D., Baumgartner, J. V., Powell, J. and Braun, D. P.
  (1996), A Method for Assessing Hydrologic Alteration within Ecosystems.
  Conservation Biology, 10: 1163-1174. doi: 10.1046/j.1523-1739.1996.10041163.x
}

