% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{IHA}
\alias{IHA}
\title{Indicators of Hydrologic Alteration}
\usage{
IHA(ts, yearstart = "01-01", yearend = "12-31", groups = 1:5, ut, lt,
  parametric = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{yearstart}{A character vector specifying the month and day signifying
the start of the water year, in the format "mm-dd". Default is "01-01".}

\item{yearend}{A character vector specifying the month and day signifying
the end of the water year, in the format "mm-dd". Default is "12-31".}

\item{groups}{The parameter groups to compute. Computes all groups (1-5) by
default.}

\item{ut}{The upper flow threshold for identifying high flow pulses. IHA 
recommends the 75th flow percentile for pre-impact conditions.}

\item{lt}{The lower flow threshold for identifying low flow pulses. IHA 
recommends the 25th flow percentile for pre-impact conditions.}

\item{parametric}{Logical: perform parametric (mean) or non-parametric
(median) analysis.}

\item{keep.raw}{Logical: return the hydrologic attributes for each year 
of record as attribute 'raw' of the output dataframe.}
}
\value{
An IHA object containing the the parameter names, the value of the
  attribute and the year of record (YoR) for which the attribute was 
  computed. Use \code{summary()} to report the central tendencies and 
  measures of dispersion for each parameter and \code{confint()} to compute
  parametric or non-parametric confidence intervals for the summary 
  statistics.
}
\description{
Compute the Indicators of Hydrologic Alteration for a daily time series.
}
\details{
The IHA method requires a regular time series with no missing 
  values.
}
\section{Notes}{

  \itemize{
    \item Low- and high-flow pulses occurring at the start of a year of 
      record are ignored when calculating number of pulses and mean duration
      of pulses.
    \item IHA v7 rounds averages of flow values to the nearest whole number 
      for individual water years, and this rounding propagates through the 
      scorecard. Numerical results may differ slightly from IHAv7.
      

}
}
\examples{
data(siouxcity)
IHA(siouxcity['2009/2011'], ut = 32000, lt = 12000)
IHA(siouxcity['2009/2011'], ut = 32000, lt = 12000, keep.raw = FALSE)
IHA(siouxcity['2009-10-01/2011-09-30'], yearstart = "10-01", 
  yearend = "09-30", ut = 32000, lt = 12000)

}
\references{
Richter, B. D., Baumgartner, J. V., Powell, J. and Braun, D. P.
  (1996), A Method for Assessing Hydrologic Alteration within Ecosystems.
  Conservation Biology, 10: 1163-1174. doi: 10.1046/j.1523-1739.1996.10041163.x
}

