% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iha.r
\name{IHAsignif}
\alias{IHAsignif}
\title{Computes significance counts for non-parametric comparison of IHA objects.}
\usage{
IHAsignif(pre, post)
}
\arguments{
\item{pre}{The pre-impact IHA statistics, i.e. the output of 
\code{IHA(...)}.}

\item{post}{The post-impact IHA statistics.}
}
\value{
A vector of deviation factors.
}
\description{
Computes significance counts for non-parametric comparison of IHA objects.
}

