% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{low_flow_duration}
\alias{low_flow_duration}
\title{Low Flow Duration}
\usage{
low_flow_duration(ts, lt, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{lt}{The lower flow threshold used to compute duration.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the 
duration instead of the count.}
}
\value{
The duration of the longest period where flow is at
  or below the threshold, in a format dependent on the value 
  of \code{index(ts)}.
}
\description{
Compute the longest continuous period during which flow is 
  at or below a given threshold.
}
\examples{
data(siouxcity)
low_flow_duration(siouxcity['2006-06/2007-06'], 18000)
low_flow_duration(siouxcity['2006-06/2007-06'], 18000, which = TRUE)

}

