% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{number_of_pulses}
\alias{number_of_pulses}
\title{Number of High Flow Pulses}
\usage{
number_of_pulses(ts, ut = 0, ws = 3, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{ut}{The upper flow threshold above which to identify peaks.}

\item{ws}{The window size within which to detect peaks. Must be odd.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the
peaks instead of the count.}
}
\value{
The number of peaks above the threshold.
}
\description{
Compute the number of high flow pulses (peaks) above a threshold.
}
\details{
The method \code{rollapply} identifies peaks by testing whether the
  value at the center of the rolling window is the maximum value within the
  window. Larger windows can be used to essentially reduce the tolerance for
  what is considered a 'peak' in the presence of noise.
}
\examples{
data(siouxcity)
number_of_pulses(siouxcity['2009'])
number_of_pulses(siouxcity['2009'], which = TRUE)
number_of_pulses(siouxcity['2009'], ws = 7, which = TRUE)
number_of_pulses(siouxcity['2009'], ws = 7, ut = 32000)
}

