% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rva.r
\name{RVA}
\alias{RVA}
\title{IHA Range of Variability Approach}
\usage{
RVA(pre, post, rvacat)
}
\arguments{
\item{pre}{The pre-impact IHA statistics, i.e. the output of 
\code{IHA(...)}.}

\item{post}{The post-impact IHA statistics.}

\item{rvacat}{The RVA categories, as defined by e.g. RVA_categories(pre)}
}
\value{
A 5-column dataframe containing the Hydrologic Alteration Factor for 
  each parameter and RVA category, and the upper and lower bounds defining 
  the RVA categories for each parameter.
}
\description{
Analyse IHA results using the Range of Variability Approach.
}
\details{
The \code{boundaries} argument defines cutoffs relative to the mean
  (if \code{parametric = TRUE}) or median (if \code{parametric = FALSE}), 
  e.g. \code{boundaries = c(-17, 17)} represents cutoffs at the 33rd and 
  67th percentile when \code{parametric = FALSE}, and 
  \code{boundaries = c(-0.5, 0.5)} represents cutoffs at +/- 0.5 standard
  deviations from the mean when \code{parametric = TRUE}.
}

