% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{time_to_recede}
\alias{time_to_recede}
\title{Time To Recede}
\usage{
time_to_recede(ts, lt, ut, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{lt}{The lower flow threshold for which the time to recede is
computed. If omitted, the minimum flow of the series is used.}

\item{ut}{The upper flow threshold from which to start computing
the time to recede. If omitted, the maximum flow of the series is
used.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the
recession limb instead of the count.}
}
\value{
The time to recede, in a format dependent on the value of
  \code{index(ts)}.
}
\description{
Compute the time for flow to fall down to or below a given threshold.
}
\examples{
data(siouxcity)
time_to_recede(siouxcity['2011'])
time_to_recede(siouxcity['2011'], which = TRUE)
time_to_recede(siouxcity['2011'], lt = 70000, which = TRUE)
}

