% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{time_to_rise}
\alias{time_to_rise}
\title{Time To Rise}
\usage{
time_to_rise(ts, ut, lt, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{ut}{The upper flow threshold for which the time to rise is
computed. If omitted, the maximum flow of the series is used.}

\item{lt}{The lower flow threshold from which to start computing
the time to rise. If omitted, the minimum flow of the series is
used.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the
rising limb instead of the count.}
}
\value{
The time to rise, in a format dependent on the value of
  \code{index(ts)}.
}
\description{
Compute the time for flow to rise up to or above a given threshold.
}
\examples{
data(siouxcity)
time_to_rise(siouxcity['2011'])
time_to_rise(siouxcity['2011'], which = TRUE)
time_to_rise(siouxcity['2011'], lt = 70000)
}

