% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowregime.R
\name{total_time_above_threshold}
\alias{total_time_above_threshold}
\title{Total Time Above Threshold}
\usage{
total_time_above_threshold(ts, ut, which = FALSE)
}
\arguments{
\item{ts}{A time series of class \code{xts}.}

\item{ut}{The upper flow threshold.}

\item{which}{Logical: If \code{TRUE}, return the index locations of the 
high flow records.}
}
\value{
The total amount of time that flow is at or 
  above the threshold, in a format dependent on the 
  value of \code{index(ts)}.
}
\description{
Compute the total amount of time that flow is 
  at or above a given threshold.
}
\examples{
data(siouxcity)
total_time_above_threshold(siouxcity['2010'], 60000)
total_time_above_threshold(siouxcity['2010'], 60000, which = TRUE)

}

