% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowregime.R
\name{total_time_above_threshold}
\alias{total_time_above_threshold}
\title{Total Time Above Threshold}
\usage{
total_time_above_threshold(ts, ut)
}
\arguments{
\item{ts}{A time series of  of class \code{xts}.}

\item{ut}{The upper flow threshold.}
}
\value{
The total amount of time that flow is at or
  above the threshold, in a format dependent on the
  value of \code{index(ts)}.
}
\description{
Compute the total amount of time that flow is
  at or above a given threshold.
}
\examples{
data(siouxcity)
upper = quantile(siouxcity, 0.75)
total_time_above_threshold(siouxcity, upper)
}

